szymonmiks-deployment/pyproject.toml
[tool.poetry]
name = "szymonmiks-deployment"
version = "0.1.0"
description = ""
authors = ["Szymon Miks <miks.szymon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.2"
paramiko = "^2.9.2"
typer = "^0.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.17.0"
black = "^22.1.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.931"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py39"]